Clean Rebuild
Cloned the previous committ, erased all contents from that dirty build,
and move the working build from my USB Drive into the current working
directory.

Created current build with Unity 4.5.2f1. If you have the newer version
of Unity3D installed (4f1 IIRC), please let me know, so that I can
upgrade my copy of Unity3D, so there will be no problems. Oppositely, if
your copy of Unity3D is older, please upgrade to the latest version.

Let me know if everything is now working, and that nothing is exploding
like before. If it is, it is either due to problems in the with .meta
files, or because you don't have the external Assets folder. If that is
the case,  I will commit those.

Most importantly, let me know if I committed everything correctly (using
GitHub and version controls is new to me.)