Powerups Effects (I), Score HUD etc
Powerup Effects:

Added effects of physical object-based powerups (blockade, portals,
bumpers, and ramps), and a stub for the gravity well (need to
find/create a script for suction/gravitational attraction.

Blockade powerup:
Consists ~8 blocks in the middle of the board on the red center line(s)
(in Battle mode, ~16 blocks in a cross-shaped arrangement). Each block
can be hit thrice, and then it is destroyed. Their position on the table
is fixed.

Bumpers:
4 bumpers position themselves randomly on the board, which act as
obstacles for the puck.  It utilizes the same function for randomizing
the powerup collectible. This function has been moved to a Position
class in the Toolbox, and has parameters in order for the space to
randomize within can be a reusable and flexible script. The bumpers
randomly plays 1 of 2 bumper sound effects when hit by the puck.

Ramps:
4 ramps randomize their position on the table. Tried to randomize the y
rotation, but was not successful. When the puck hit or rides up its
slopes, the rigidbody restrictions on it are removed (it can now move on
the y axis, and gravity is enabled). When the puck collides with an
object with a Tag of "Ground", the restrictions are enabled again.

Portals:
One red and one blue portal are randomized on the board. A puck
colliding with one portal spits it out from the other. The teleportation
position is 5 units above the portal, the y axis movement restriction
becomes removed on teleportation, and gravity is enabled (until
collision with the player field enables the restrictions again)

Gravity Well:
The gravity well is supposed to attract  pucks towards it, but the
script is simply a stub for now.

All powerups:
All object-based powerups are supposed to have the gameobjects disabled
upon game init. An invisible game thinker object will handle the powerup
usage, powerup processing, and powerup collection. This object and its
script needs to be written by next week. This object will enable/disable
the appropriate powerup objects via a Tag.
-----

Score HUD etc:
Added a Score HUD. This HUD consists of 4 Text Meshes, one for each
color, and is placed over what used to a decoration on the score box
above the middle of the table. Each single-digit score has the
appropriate player's colors, and updates as appropriate. Each side of
the box has the 4 digits

When the game is complete, these digits flash character digit words,
displaying either "Game Over [Some team color] Wins" or "Game Over Tie".
This feature is not fully functional yet. (Need to make a delay in
switching the display of each word.)

Powerup HUD:
Created simple powerup HUD. Need to create script to process it. It
contains a box, and a text mesh. Box will change its texture to the
powerup selected. Upon clicking left/right mouse button, will be able to
cycle through the powerup(s) held in reserve, and the be able to use the
powerup. Text mesh shows the multiplicity of the powerup.

Paddle/Pucks:
Changed R/B/G/Y colors to have self-illumination, in order to have the
neon effect our MASD students wanted

Playerfields:
Sub divided the playerfield planes. Still have R/G/B/Y fields, but they
are only near the goal. Add additional playerfields in with two color
tags for the overlapping game play areas. Bug fixed the code so that
each paddle will only be able to raycast the paddle in its appropriate
player field zones.

Puck respawn:

Fixed and re-enabled puck respawning. Now works by moving puck to
PuckRespawn upon hitting an object with the "Reset" tag (Goals, out of
bound box detectors, &c). Puck then has movement restrictions removed.
Puck then falls onto ground, and has those restriction enabled again.
-----------------------

Any problems, email me. Do to you only working on the Options and
Credits at the moment, you should be able to simply drag and drop your
changes onto mine. (I only made changes to the Game scene, added some
scripts and objects, and modified the Toolbox script).

"Tamk1s"